+++ src/__experimental__/components/date/date-picker.component.js
@@ -45,8 +45,8 @@ const DatePicker = (props) => {
     selectedDays: props.selectedDate || undefined,
     weekdayElement: (weekdayElementProps) => {
       const { className, weekday, localeUtils } = weekdayElementProps;
-      const weekdayLong = localeUtils.formatWeekdayLong(weekday);
-      const weekdayShort = weekdayLong.substring(0, 3);
+      const weekdayLong = localeUtils.formatWeekdayLong(weekday, I18n.locale);
+      const weekdayShort = localeUtils.formatWeekdayShort(weekday, I18n.locale);
 
       return (
         <Weekday className={ className } title={ weekdayLong }>